Fixed silly arguments processing
The Teachers JUnit tests required to do a stupid arguments processing and to make the user to select an interface an a text file with the map mandatory.
I've fixed since there is no need to be like that anymore, now it does better checking and it has default values for those options.

 On branch master
 Changes to be committed:
	modified:   README.md
	modified:   src/tp/pr5/Main.java